+++ app/uk/gov/hmrc/individualsifapistub/services/IncomeService.scala
@@ -17,10 +17,9 @@
 package uk.gov.hmrc.individualsifapistub.services
 
 import javax.inject.Inject
-import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.http.{BadRequestException, HeaderCarrier}
+import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.individualsifapistub.connector.ApiPlatformTestUserConnector
-import uk.gov.hmrc.individualsifapistub.domain.{IdType, IncomePaye, IncomeSa, RecordNotFoundException}
+import uk.gov.hmrc.individualsifapistub.domain.{IncomePaye, IncomeSa}
 import uk.gov.hmrc.individualsifapistub.repository.{IncomePayeRepository, IncomeSaRepository}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 
@@ -29,7 +28,7 @@ import scala.concurrent.{ExecutionContext, Future}
 class IncomeService @Inject()(incomeSaRepository: IncomeSaRepository,
                               incomePayeRepository: IncomePayeRepository,
                               apiPlatformTestUserConnector: ApiPlatformTestUserConnector,
-                              servicesConfig: ServicesConfig) {
+                              servicesConfig: ServicesConfig) extends ServiceBase(apiPlatformTestUserConnector) {
 
   def createSa(idType: String,
                idValue: String,
@@ -72,19 +71,5 @@ class IncomeService @Inject()(incomeSaRepository: IncomeSaRepository,
               endDate: String,
               fields: Option[String]): Future[Option[IncomePaye]] = {
     incomePayeRepository.findByTypeAndId(idType, idValue, startDate, endDate, fields)
-  }
-
-  def verifyNino(idType: String, idValue: String)
-                (implicit ec: ExecutionContext,
-                 hc: HeaderCarrier) = {
-    IdType.parse(idType) match {
-      case IdType.Nino => {
-        for {
-          individual <- apiPlatformTestUserConnector.getIndividualByNino(Nino(idValue))
-          utr = individual.saUtr.getOrElse(throw new RecordNotFoundException)
-        } yield utr
-      }
-      case _ => throw new BadRequestException("Invalid National Insurance Number")
-    }
   }
 }
