+++ app/uk/gov/hmrc/individualsifapistub/services/TaxCreditsService.scala
@@ -17,10 +17,9 @@
 package uk.gov.hmrc.individualsifapistub.services
 
 import javax.inject.Inject
-import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.http.{BadRequestException, HeaderCarrier}
+import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.individualsifapistub.connector.ApiPlatformTestUserConnector
-import uk.gov.hmrc.individualsifapistub.domain.{Applications, IdType, RecordNotFoundException}
+import uk.gov.hmrc.individualsifapistub.domain.Applications
 import uk.gov.hmrc.individualsifapistub.repository.TaxCreditsRepository
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 
@@ -28,7 +27,7 @@ import scala.concurrent.{ExecutionContext, Future}
 
 class TaxCreditsService @Inject()(repository: TaxCreditsRepository,
                                   apiPlatformTestUserConnector: ApiPlatformTestUserConnector,
-                                  servicesConfig: ServicesConfig) {
+                                  servicesConfig: ServicesConfig) extends ServiceBase(apiPlatformTestUserConnector) {
 
   def create(idType: String,
              idValue: String,
@@ -49,19 +48,5 @@ class TaxCreditsService @Inject()(repository: TaxCreditsRepository,
           endDate: String,
           fields: Option[String]): Future[Option[Applications]] = {
     repository.findByIdAndType(idType, idValue, startDate, endDate, fields)
-  }
-
-  def verifyNino(idType: String, idValue: String)
-                (implicit ec: ExecutionContext,
-                 hc: HeaderCarrier) = {
-    IdType.parse(idType) match {
-      case IdType.Nino => {
-        for {
-          individual <- apiPlatformTestUserConnector.getIndividualByNino(Nino(idValue))
-          utr = individual.saUtr.getOrElse(throw new RecordNotFoundException)
-        } yield utr
-      }
-      case _ => throw new BadRequestException("Invalid National Insurance Number")
-    }
   }
 }
