+++ test/it/uk/gov/hmrc/individualsifapistub/EmploymentRepositorySpec.scala
@@ -80,8 +80,8 @@ class EmploymentRepositorySpec extends RepositoryTestHelper {
 
       await(repository.collection.indexesManager.list()).find({ i =>
       {
-        i.name.contains("cache-key") &&
-          i.key.exists(key => key._1 == "details")
+        i.name.contains("id") &&
+          i.key.exists(key => key._1 == "id")
         i.background &&
           i.unique
       }
